changed dashboard to class component in order to fetch companies on it's mount
